+++ plugins/baserobots/baserobots.cpp
@@ -1,5 +1,5 @@
 // -*- coding: utf-8 --*
-// Copyright (C) 2006-2011 Rosen Diankov <rosen.diankov@gmail.com>
+// Copyright (C) 2006-2012 Rosen Diankov <rosen.diankov@gmail.com>
 //
 // This file is part of OpenRAVE.
 // OpenRAVE is free software: you can redistribute it and/or modify
@@ -15,37 +15,49 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "plugindefs.h"
-#include "collisionmaprobot.h"
 #include <openrave/plugin.h>
 
 static UserDataPtr s_RegisteredReader;
 
+RobotBasePtr CreateCollisionMapRobot(EnvironmentBasePtr penv, std::istream& sinput);
+RobotBasePtr CreateConveyorRobot(EnvironmentBasePtr penv, std::istream& sinput);
+
+void RegisterCollisionMapRobotReaders(std::list< UserDataPtr >& listRegisteredReaders);
+void RegisterConveyorReaders(std::list< UserDataPtr >& listRegisteredReaders);
+
+static std::list< UserDataPtr >* s_listRegisteredReaders = NULL; ///< have to make it a pointer in order to prevent static object destruction from taking precedence
+
 InterfaceBasePtr CreateInterfaceValidated(InterfaceType type, const std::string& interfacename, std::istream& sinput, EnvironmentBasePtr penv)
 {
-    if( !s_RegisteredReader ) {
-        /// as long as this pointer is valid, the reader will remain registered
-        s_RegisteredReader = RaveRegisterXMLReader(PT_Robot,"collisionmap",CollisionMapRobot::CreateXMLReader);
+    if( !s_listRegisteredReaders ) {
+        s_listRegisteredReaders = new list< UserDataPtr >();
+        RegisterCollisionMapRobotReaders(*s_listRegisteredReaders);
+        RegisterConveyorReaders(*s_listRegisteredReaders);
     }
 
     switch(type) {
     case PT_Robot:
         if( interfacename == "collisionmaprobot") {
-            return InterfaceBasePtr(new CollisionMapRobot(penv));
+            return CreateCollisionMapRobot(penv,sinput);
+        }
+        else if( interfacename == "conveyor" ) {
+            return CreateConveyorRobot(penv,sinput);
         }
         break;
     default:
         break;
     }
-
     return InterfaceBasePtr();
 }
 
 void GetPluginAttributesValidated(PLUGININFO& info)
 {
     info.interfacenames[PT_Robot].push_back("CollisionMapRobot");
+    info.interfacenames[PT_Robot].push_back("Conveyor");
 }
 
 OPENRAVE_PLUGIN_API void DestroyPlugin()
 {
-    s_RegisteredReader.reset(); // unregister the reader
+    delete s_listRegisteredReaders;
+    s_listRegisteredReaders = NULL;
 }